# Entrega Final/src/docs/swagger.yaml
openapi: 3.0.3

info:
  title: Pet Adoption API
  description: Full professional API documentation with reusable responses, schemas and security
  version: 1.0.0

servers:
  - url: http://localhost:8080

tags:
  - name: Root
  - name: Users
  - name: Pets
  - name: Adoptions
  - name: Mocks

components:

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:

    BaseResponse:
      type: object
      properties:
        status:
          type: string
          example: success

    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          example: Error message

    User:
      type: object
      properties:
        _id:
          type: string
          example: 64f8a123456789abcdef1234
        first_name:
          type: string
          example: Facundo
        last_name:
          type: string
          example: Lodi
        email:
          type: string
          example: lodifacundo3@mail.com
        age:
          type: number
          example: 24
        role:
          type: string
          example: user
        pets:
          type: array
          items:
            type: string

    Pet:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
          example: Beky
        specie:
          type: string
          example: cat
        birthDate:
          type: string
          format: date
        adopted:
          type: boolean
          example: false

    Adoption:
      type: object
      properties:
        _id:
          type: string
        user:
          type: object
          properties:
            _id:
              type: string
            email:
              type: string
        pet:
          type: object
          properties:
            _id:
              type: string
            name:
              type: string
        adoptedAt:
          type: string
          format: date

    CreateUser:
      type: object
      required: [first_name, last_name, email, password]
      properties:
        first_name:
          type: string
          example: Facundo
        last_name:
          type: string
          example: Lodi
        email:
          type: string
          example: lodifacundo3@mail.com
        age:
          type: number
          example: 24
        password:
          type: string
          example: 123456

    CreatePet:
      type: object
      required: [name, specie, birthDate]
      properties:
        name:
          type: string
          example: Beky
        specie:
          type: string
          example: cat
        birthDate:
          type: string
          format: date

    CreateAdoption:
      type: object
      required: [userId, petId]
      properties:
        userId:
          type: string
        petId:
          type: string

    GenerateMockData:
      type: object
      properties:
        users:
          type: number
          example: 10
        pets:
          type: number
          example: 10

  responses:

    Success:
      description: Successful operation
      content:
        application/json:
          example:
            status: success

    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

paths:

  /:
    get:
      tags: [Root]
      summary: Health check
      responses:
        200:
          description: API running
          content:
            application/json:
              example:
                status: success
                message: Pet Adoption API running

  /api/users:
    get:
      tags: [Users]
      summary: Get all users
      responses:
        200:
          description: Users retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  payload:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        500:
          $ref: '#/components/responses/ServerError'

    post:
      tags: [Users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        201:
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  payload:
                    $ref: '#/components/schemas/User'
        400:
          $ref: '#/components/responses/BadRequest'
        500:
          $ref: '#/components/responses/ServerError'

  /api/users/{uid}:
    get:
      tags: [Users]
      summary: Get user by ID
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: User found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  payload:
                    $ref: '#/components/schemas/User'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Users]
      summary: Delete user
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: User deleted
          content:
            application/json:
              example:
                status: success
                message: Usuario eliminado correctamente
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'

  /api/pets:
    get:
      tags: [Pets]
      summary: Get all pets
      responses:
        200:
          description: Pets list
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  payload:
                    type: array
                    items:
                      $ref: '#/components/schemas/Pet'
        500:
          $ref: '#/components/responses/ServerError'

    post:
      tags: [Pets]
      summary: Create pet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePet'
      responses:
        201:
          description: Pet created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  payload:
                    $ref: '#/components/schemas/Pet'
        400:
          $ref: '#/components/responses/BadRequest'

  /api/pets/{pid}:
    get:
      tags: [Pets]
      summary: Get pet by ID
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Pet found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  payload:
                    $ref: '#/components/schemas/Pet'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Pets]
      summary: Update pet
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePet'
      responses:
        200:
          description: Pet updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  payload:
                    $ref: '#/components/schemas/Pet'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Pets]
      summary: Delete pet
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Pet deleted
          content:
            application/json:
              example:
                status: success
                message: Mascota eliminada correctamente
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'

  /api/adoptions:
    get:
      tags: [Adoptions]
      summary: Get all adoptions
      responses:
        200:
          description: Adoptions list
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  payload:
                    type: array
                    items:
                      $ref: '#/components/schemas/Adoption'
        500:
          $ref: '#/components/responses/ServerError'

    post:
      tags: [Adoptions]
      summary: Create adoption
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAdoption'
      responses:
        201:
          description: Adoption created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  payload:
                    $ref: '#/components/schemas/Adoption'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'

  /api/adoptions/{aid}:
    get:
      tags: [Adoptions]
      summary: Get adoption by ID
      parameters:
        - name: aid
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Adoption found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  payload:
                    $ref: '#/components/schemas/Adoption'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'

  /api/mocks/mockingpets:
    get:
      tags: [Mocks]
      summary: Generate mock pets
      responses:
        200:
          description: Mock pets generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  payload:
                    type: array
                    items:
                      $ref: '#/components/schemas/Pet'

  /api/mocks/mockingusers:
    get:
      tags: [Mocks]
      summary: Generate mock users
      responses:
        200:
          description: Mock users generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  payload:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'

  /api/mocks/generateData:
    post:
      tags: [Mocks]
      summary: Insert mock data into DB
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateMockData'
      responses:
        200:
          description: Data inserted
          content:
            application/json:
              example:
                status: success
                message: Inserted X users and X pets
        500:
          $ref: '#/components/responses/ServerError'